Complete test for 'to_constant' function factory
