+++ tests/functional_tests/to_constant_tests/test_pickle.py
@@ -1,0 +1,20 @@
+from typing import (Any,
+                    Dict,
+                    Tuple)
+
+from lz.functional import to_constant
+from tests.utils import (are_objects_similar,
+                         round_trip_pickle)
+
+
+def test_round_trip(object_: Any,
+                    positional_arguments: Tuple,
+                    keyword_arguments: Dict[str, Any]) -> None:
+    constant = to_constant(object_)
+
+    result = round_trip_pickle(constant)
+
+    assert are_objects_similar(result(*positional_arguments,
+                                      **keyword_arguments),
+                               constant(*positional_arguments,
+                                        **keyword_arguments))
